Add JavaDocs for commands and fix style
