+++ src/main/java/commands/FindCommand.java
@@ -1,16 +1,36 @@
 package commands;
+
 import storage.Storage;
 import storage.TaskList;
 import tasks.Task;
 import ui.UI;
 
+/**
+ * Represents a command to find tasks that contain a specific keyword or phrase.
+ * The FindCommand class searches through the task list for tasks that match the specified description.
+ */
 public class FindCommand implements Command {
     private final String desc;
 
+    /**
+     * Constructs a FindCommand with the specified keyword or phrase to search for.
+     *
+     * @param desc the keyword or phrase to search for in the task descriptions
+     */
     public FindCommand(String desc) {
         this.desc = desc;
     }
 
+    /**
+     * Executes the find command, searching for tasks in the task list that contain the specified keyword or phrase.
+     * If matching tasks are found, they are displayed to the user.
+     * If no matches are found, an appropriate message is shown.
+     *
+     * @param storage the Storage object for handling task persistence
+     * @param master the TaskList object containing the list of tasks
+     * @param ui the UI object for interacting with the user
+     * @return false, indicating that the application should not terminate
+     */
     @Override
     public boolean execute(Storage storage, TaskList master, UI ui) {
         String s = "";
